+++ packages/keyring-api/src/api/index.ts
@@ -8,4 +8,5 @@ export * from './request';
 export * from './response';
 export * from './transaction';
 export * from './pagination';
+export * from './options';
 export type * from './keyring';