+++ packages/keyring-eth-ledger-bridge/src/ledger-keyring.test.ts
@@ -3,7 +3,7 @@ import { RLP } from '@ethereumjs/rlp';
 import { TransactionFactory } from '@ethereumjs/tx';
 import * as ethUtil from '@ethereumjs/util';
 import * as sigUtil from '@metamask/eth-sig-util';
-import { bytesToHex, Hex } from '@metamask/utils';
+import { bytesToHex, Hex, remove0x } from '@metamask/utils';
 import EthereumTx from 'ethereumjs-tx';
 import HDKey from 'hdkey';
 
@@ -719,7 +719,9 @@ describe('LedgerKeyring', function () {
             .mockImplementation(async (params) => {
               expect(params).toStrictEqual({
                 hdPath: "m/44'/60'/0'/0",
-                tx: bytesToHex(fakeTypeTwoTx.getMessageToSign() as Uint8Array),
+                tx: remove0x(
+                  bytesToHex(fakeTypeTwoTx.getMessageToSign() as Uint8Array),
+                ),
               });
               return expectedRSV;
             });
@@ -745,7 +747,7 @@ describe('LedgerKeyring', function () {
 
         await expect(
           keyring.signTransaction(fakeAccounts[0], fakeTx),
-        ).rejects.toThrow('Ledger: Unknown error while signing transaction');
+        ).rejects.toThrow('Ledger: hdPath is empty while signing transaction');
       });
 
       it('throws different error to the default one if the bridge error is an instance of the Error object', async function () {
@@ -968,7 +970,6 @@ describe('LedgerKeyring', function () {
           ],
         },
       };
-      const options = { version: 'V4' };
 
       beforeEach(async function () {
         jest
@@ -989,7 +990,7 @@ describe('LedgerKeyring', function () {
         const result = await keyring.signTypedData(
           fakeAccounts[15],
           fixtureData,
-          options,
+          { version: sigUtil.SignTypedDataVersion.V4 },
         );
         expect(result).toBe(
           '0x72d4e38a0e582e09a620fd38e236fe687a1ec782206b56d576f579c026a7e5b946759735981cd0c3efb02d36df28bb2feedfec3d90e408efc93f45b894946e321b',
@@ -1016,7 +1017,7 @@ describe('LedgerKeyring', function () {
         const result = await keyring.signTypedData(
           fakeAccounts[15],
           fixtureDataWithoutSalt,
-          options,
+          { version: sigUtil.SignTypedDataVersion.V4 },
         );
         expect(result).toBe(
           '0x72d4e38a0e582e09a620fd38e236fe687a1ec782206b56d576f579c026a7e5b946759735981cd0c3efb02d36df28bb2feedfec3d90e408efc93f45b894946e321b',
@@ -1034,7 +1035,9 @@ describe('LedgerKeyring', function () {
           }));
 
         await expect(
-          keyring.signTypedData(fakeAccounts[15], fixtureData, options),
+          keyring.signTypedData(fakeAccounts[15], fixtureData, {
+            version: sigUtil.SignTypedDataVersion.V4,
+          }),
         ).rejects.toThrow(
           'Ledger: The signature doesnt match the right address',
         );
@@ -1043,7 +1046,8 @@ describe('LedgerKeyring', function () {
       it('throws an error if the signTypedData version is not v4', async function () {
         await expect(
           keyring.signTypedData(fakeAccounts[0], fixtureData, {
-            version: 'V3',
+            // @ts-expect-error we want to test an invalid version
+            version: sigUtil.SignTypedDataVersion.V3,
           }),
         ).rejects.toThrow(
           'Ledger: Only version 4 of typed data signing is supported',
@@ -1063,7 +1067,9 @@ describe('LedgerKeyring', function () {
           .spyOn(keyring, 'unlockAccountByAddress')
           .mockResolvedValue(undefined);
         await expect(
-          keyring.signTypedData(fakeAccounts[0], fixtureData, options),
+          keyring.signTypedData(fakeAccounts[0], fixtureData, {
+            version: sigUtil.SignTypedDataVersion.V4,
+          }),
         ).rejects.toThrow('Ledger: Unknown error while signing message');
       });
 
@@ -1073,7 +1079,9 @@ describe('LedgerKeyring', function () {
           .mockRejectedValue(new Error('some error'));
 
         await expect(
-          keyring.signTypedData(fakeAccounts[15], fixtureData, options),
+          keyring.signTypedData(fakeAccounts[15], fixtureData, {
+            version: sigUtil.SignTypedDataVersion.V4,
+          }),
         ).rejects.toThrow('some error');
       });
 
@@ -1083,7 +1091,9 @@ describe('LedgerKeyring', function () {
           .mockRejectedValue('some error');
 
         await expect(
-          keyring.signTypedData(fakeAccounts[15], fixtureData, options),
+          keyring.signTypedData(fakeAccounts[15], fixtureData, {
+            version: sigUtil.SignTypedDataVersion.V4,
+          }),
         ).rejects.toThrow('Ledger: Unknown error while signing message');
       });
 
@@ -1096,7 +1106,9 @@ describe('LedgerKeyring', function () {
         const result = await keyring.signTypedData(
           fakeAccounts[15],
           fixtureData,
-          options,
+          {
+            version: sigUtil.SignTypedDataVersion.V4,
+          },
         );
         expect(result).toBe(
           '0x72d4e38a0e582e09a620fd38e236fe687a1ec782206b56d576f579c026a7e5b946759735981cd0c3efb02d36df28bb2feedfec3d90e408efc93f45b894946e3200',